fix(auth): fix register method error because of unmatching entities

+++ app/infrastructure/db/repositories/activation_token_repository.py
@@ -1,7 +1,10 @@
 from datetime import datetime
 from sqlalchemy.orm import Session
+
 from app.domain.entities.activation_token import ActivationToken
-from app.infrastructure.db.mappers.activation_token_mapper import to_model
+from app.infrastructure.db.models.activation_token import ActivationToken as ActivationTokenModel
+from app.infrastructure.db.mappers.activation_token_mapper import to_model, to_domain
+
 
 class ActivationTokenRepository:
 
@@ -14,16 +17,22 @@ class ActivationTokenRepository:
         self.db.commit()
 
     def get_valid_token(self, token_str: str) -> ActivationToken | None:
-        return (
-            self.db.query(ActivationToken)
+        model = (
+            self.db.query(ActivationTokenModel)
             .filter(
-                ActivationToken.token == token_str,
-                ActivationToken.expires_at > datetime.utcnow()
+                ActivationTokenModel.token == token_str,
+                ActivationTokenModel.expires_at > datetime.utcnow()
             )
             .first()
         )
+        return to_domain(model) if model else None
 
     def delete(self, token: ActivationToken):
-        model = to_model(token)
-        self.db.delete(model)
-        self.db.commit()
+        model = (
+            self.db.query(ActivationTokenModel)
+            .filter(ActivationTokenModel.token == token.token)
+            .first()
+        )
+        if model:
+            self.db.delete(model)
+            self.db.commit()
